# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5f46db3167541c707d179879
  title: appvaa
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/clubs/{id}":
    delete:
      x-skaffolder-id: 5f46ead567541c707d17999b
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5f46e51a67541c707d179939
      x-skaffolder-resource: Clubs
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Clubs
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5f46ead567541c707d17998b
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5f46e51a67541c707d179939
      x-skaffolder-resource: Clubs
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Clubs
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clubs"
    post:
      x-skaffolder-id: 5f46ead567541c707d17998d
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5f46e51a67541c707d179939
      x-skaffolder-resource: Clubs
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Clubs
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clubs"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Clubs"
  "/clubs/findBy_class/{key}":
    get:
      x-skaffolder-id: 5f46eb1867541c707d1799cb
      x-skaffolder-name: findBy_class
      x-skaffolder-id-resource: 5f46e51a67541c707d179939
      x-skaffolder-resource: Clubs
      x-skaffolder-crudAction: findBy_class
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_class
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_class/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_class
      tags:
        - Clubs
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clubs"
  "/clubs/findBy_teacher/{key}":
    get:
      x-skaffolder-id: 5f46eac167541c707d179978
      x-skaffolder-name: findBy_teacher
      x-skaffolder-id-resource: 5f46e51a67541c707d179939
      x-skaffolder-resource: Clubs
      x-skaffolder-crudAction: findBy_teacher
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_teacher
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_teacher/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_teacher
      tags:
        - Clubs
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clubs"
...
  "/countries/{id}":
    delete:
      x-skaffolder-id: 5f46eaf067541c707d1799af
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5f46e54267541c707d179944
      x-skaffolder-resource: Countries
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Countries
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5f46eaef67541c707d1799aa
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5f46e54267541c707d179944
      x-skaffolder-resource: Countries
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Countries
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Countries"
    post:
      x-skaffolder-id: 5f46eaef67541c707d1799a8
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5f46e54267541c707d179944
      x-skaffolder-resource: Countries
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Countries
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Countries"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Countries"
...
  "/events/{id}":
    delete:
      x-skaffolder-id: 5f46eae167541c707d1799a4
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5f46e53567541c707d17993f
      x-skaffolder-resource: Events
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Events
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5f46eae167541c707d17999f
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5f46e53567541c707d17993f
      x-skaffolder-resource: Events
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Events
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Events"
    post:
      x-skaffolder-id: 5f46eae167541c707d1799a0
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5f46e53567541c707d17993f
      x-skaffolder-resource: Events
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Events
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Events"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Events"
  "/events/findBy_club/{key}":
    get:
      x-skaffolder-id: 5f46ead567541c707d179994
      x-skaffolder-name: findBy_club
      x-skaffolder-id-resource: 5f46e53567541c707d17993f
      x-skaffolder-resource: Events
      x-skaffolder-crudAction: findBy_club
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_club
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_club/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_club
      tags:
        - Events
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Events"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Clubs:
      x-skaffolder-id: 5f46e51a67541c707d179939
      x-skaffolder-id-db: 5f46db3167541c707d17987a
      x-skaffolder-id-entity: 5f46e51a67541c707d17992b
      x-skaffolder-url: /clubs
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        canoe_oc1:
          type: number
          x-skaffolder-id-attr: 5f46e51a67541c707d17992c
          x-skaffolder-type: Number
        canoe_oc2:
          type: number
          x-skaffolder-id-attr: 5f46e51a67541c707d17992d
          x-skaffolder-type: Number
        canoe_oc6:
          type: number
          x-skaffolder-id-attr: 5f46e51a67541c707d17992e
          x-skaffolder-type: Number
        canoe_oc6_7:
          type: number
          x-skaffolder-id-attr: 5f46e51a67541c707d17992f
          x-skaffolder-type: Number
        canoe_v1:
          type: number
          x-skaffolder-id-attr: 5f46e51a67541c707d179930
          x-skaffolder-type: Number
        canoe_v3:
          type: number
          x-skaffolder-id-attr: 5f46e51a67541c707d179931
          x-skaffolder-type: Number
        canoe_v6:
          type: number
          x-skaffolder-id-attr: 5f46e51a67541c707d179932
          x-skaffolder-type: Number
        city:
          type: string
          x-skaffolder-id-attr: 5f46e51a67541c707d179933
          x-skaffolder-type: String
          x-skaffolder-required: true
        country:
          type: string
          x-skaffolder-id-attr: 5f46e51a67541c707d179934
          x-skaffolder-type: String
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5f46e51a67541c707d179935
          x-skaffolder-type: String
          x-skaffolder-required: true
        schedule:
          type: number
          x-skaffolder-id-attr: 5f46e51a67541c707d179936
          x-skaffolder-type: Number
          x-skaffolder-required: true
        state:
          type: string
          x-skaffolder-id-attr: 5f46e51a67541c707d179937
          x-skaffolder-type: String
          x-skaffolder-required: true
        street:
          type: string
          x-skaffolder-id-attr: 5f46e51a67541c707d179938
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations:
        _class:
          x-skaffolder-id: 5f46e5eb67541c707d179959
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5f46e51a67541c707d17992b
          x-skaffolder-ent2: 5f46e59767541c707d179951
        _teacher:
          x-skaffolder-id: 5f46e5eb67541c707d17995a
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5f46e51a67541c707d17992b
          x-skaffolder-ent2: 5f46e3f967541c707d17990c
      required:
        - _id
        - city
        - country
        - name
        - schedule
        - state
        - street
    Countries:
      x-skaffolder-id: 5f46e54267541c707d179944
      x-skaffolder-id-db: 5f46db3167541c707d17987a
      x-skaffolder-id-entity: 5f46e54267541c707d179942
      x-skaffolder-url: /countries
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5f46e54267541c707d179943
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - name
    Events:
      x-skaffolder-id: 5f46e53567541c707d17993f
      x-skaffolder-id-db: 5f46db3167541c707d17987a
      x-skaffolder-id-entity: 5f46e53467541c707d17993c
      x-skaffolder-url: /events
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5f46e53467541c707d17993d
          x-skaffolder-type: String
          x-skaffolder-required: true
        vacancies:
          type: string
          x-skaffolder-id-attr: 5f46e53467541c707d17993e
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations:
        _club:
          x-skaffolder-id: 5f46e5b667541c707d179958
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5f46e53467541c707d17993c
          x-skaffolder-ent2: 5f46e51a67541c707d17992b
      required:
        - _id
        - name
        - vacancies
...